Fix file linting with Stylelint
Change linting with Stylelint. Now all the dependencies it uses
are added explicitly, instead of packed into 1 package (previously there
was a plugin stylelint-config-recommended-scss).

Change the file ignoring so that Stylelint, Prettier, and Eslint only
check the correct folders.

Add a stylelint-webpack-plugin to link style files when building
a project.